Sincroniza cambios locales backend

+++ src/main/java/org/example/repositories/BonoRepository.java
@@ -3,5 +3,5 @@ package org.example.repositories;
 import org.example.entities.BonoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BonoRepository extends JpaRepository<BonoEntity, Float> {
+public interface BonoRepository extends JpaRepository<BonoEntity, Long> {
 }
